# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Your App API Documentation'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://homzen.test'
...
paths:
...
  /api/v1/features:
    get:
      summary: 'List features'
      operationId: listFeatures
      description: ''
      parameters:
        -
          in: query
          name: page
          description: 'Current page of the collection. Default: 1'
          example: 17
          required: false
          schema:
            type: integer
            description: 'Current page of the collection. Default: 1'
            example: 17
            nullable: false
        -
          in: query
          name: per_page
          description: 'Maximum number of items to be returned in result set. Default: 10'
          example: 17
          required: false
          schema:
            type: integer
            description: 'Maximum number of items to be returned in result set. Default: 10'
            example: 17
            nullable: false
        -
          in: query
          name: search
          description: 'Limit results to those matching a string.'
          example: consequatur
          required: false
          schema:
            type: string
            description: 'Limit results to those matching a string.'
            example: consequatur
            nullable: false
        -
          in: query
          name: order
          description: 'Order sort attribute ascending or descending. Default: asc. One of: asc, desc'
          example: consequatur
          required: false
          schema:
            type: string
            description: 'Order sort attribute ascending or descending. Default: asc. One of: asc, desc'
            example: consequatur
            nullable: false
        -
          in: query
          name: order_by
          description: 'Sort collection by object attribute. Default: name. One of: created_at, updated_at, name'
          example: consequatur
          required: false
          schema:
            type: string
            description: 'Sort collection by object attribute. Default: name. One of: created_at, updated_at, name'
            example: consequatur
            nullable: false
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data: []
                  links:
                    first: 'http://homzen.test/api/v1/features?page=1'
                    last: 'http://homzen.test/api/v1/features?page=1'
                    prev: 'http://homzen.test/api/v1/features?page=16'
                    next: null
                  meta:
                    current_page: 17
                    from: null
                    last_page: 1
                    links:
                      -
                        url: 'http://homzen.test/api/v1/features?page=16'
                        label: '&laquo; Previous'
                        active: false
                      -
                        url: 'http://homzen.test/api/v1/features?page=1'
                        label: '1'
                        active: false
                      -
                        url: null
                        label: 'Next &raquo;'
                        active: false
                    path: 'http://homzen.test/api/v1/features'
                    per_page: 17
                    to: null
                    total: 12
                  error: false
                  message: null
                properties:
                  data:
                    type: array
                    example: []
                  links:
                    type: object
                    properties:
                      first:
                        type: string
                        example: 'http://homzen.test/api/v1/features?page=1'
                      last:
                        type: string
                        example: 'http://homzen.test/api/v1/features?page=1'
                      prev:
                        type: string
                        example: 'http://homzen.test/api/v1/features?page=16'
                      next:
                        type: string
                        example: null
                  meta:
                    type: object
                    properties:
                      current_page:
                        type: integer
                        example: 17
                      from:
                        type: string
                        example: null
                      last_page:
                        type: integer
                        example: 1
                      links:
                        type: array
                        example:
                          -
                            url: 'http://homzen.test/api/v1/features?page=16'
                            label: '&laquo; Previous'
                            active: false
                          -
                            url: 'http://homzen.test/api/v1/features?page=1'
                            label: '1'
                            active: false
                          -
                            url: null
                            label: 'Next &raquo;'
                            active: false
                        items:
                          type: object
                          properties:
                            url:
                              type: string
                              example: 'http://homzen.test/api/v1/features?page=16'
                            label:
                              type: string
                              example: '&laquo; Previous'
                            active:
                              type: boolean
                              example: false
                      path:
                        type: string
                        example: 'http://homzen.test/api/v1/features'
                      per_page:
                        type: integer
                        example: 17
                      to:
                        type: string
                        example: null
                      total:
                        type: integer
                        example: 12
                  error:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: null
      tags:
        - 'Real Estate'
      security: []
...
  /api/v1/facilities:
    get:
      summary: 'List facilities'
      operationId: listFacilities
      description: ''
      parameters:
        -
          in: query
          name: page
          description: 'Current page of the collection. Default: 1'
          example: 17
          required: false
          schema:
            type: integer
            description: 'Current page of the collection. Default: 1'
            example: 17
            nullable: false
        -
          in: query
          name: per_page
          description: 'Maximum number of items to be returned in result set. Default: 10'
          example: 17
          required: false
          schema:
            type: integer
            description: 'Maximum number of items to be returned in result set. Default: 10'
            example: 17
            nullable: false
        -
          in: query
          name: search
          description: 'Limit results to those matching a string.'
          example: consequatur
          required: false
          schema:
            type: string
            description: 'Limit results to those matching a string.'
            example: consequatur
            nullable: false
        -
          in: query
          name: order
          description: 'Order sort attribute ascending or descending. Default: asc. One of: asc, desc'
          example: consequatur
          required: false
          schema:
            type: string
            description: 'Order sort attribute ascending or descending. Default: asc. One of: asc, desc'
            example: consequatur
            nullable: false
        -
          in: query
          name: order_by
          description: 'Sort collection by object attribute. Default: name. One of: created_at, updated_at, name'
          example: consequatur
          required: false
          schema:
            type: string
            description: 'Sort collection by object attribute. Default: name. One of: created_at, updated_at, name'
            example: consequatur
            nullable: false
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data: []
                  links:
                    first: 'http://homzen.test/api/v1/facilities?page=1'
                    last: 'http://homzen.test/api/v1/facilities?page=1'
                    prev: 'http://homzen.test/api/v1/facilities?page=16'
                    next: null
                  meta:
                    current_page: 17
                    from: null
                    last_page: 1
                    links:
                      -
                        url: 'http://homzen.test/api/v1/facilities?page=16'
                        label: '&laquo; Previous'
                        active: false
                      -
                        url: 'http://homzen.test/api/v1/facilities?page=1'
                        label: '1'
                        active: false
                      -
                        url: null
                        label: 'Next &raquo;'
                        active: false
                    path: 'http://homzen.test/api/v1/facilities'
                    per_page: 17
                    to: null
                    total: 11
                  error: false
                  message: null
                properties:
                  data:
                    type: array
                    example: []
                  links:
                    type: object
                    properties:
                      first:
                        type: string
                        example: 'http://homzen.test/api/v1/facilities?page=1'
                      last:
                        type: string
                        example: 'http://homzen.test/api/v1/facilities?page=1'
                      prev:
                        type: string
                        example: 'http://homzen.test/api/v1/facilities?page=16'
                      next:
                        type: string
                        example: null
                  meta:
                    type: object
                    properties:
                      current_page:
                        type: integer
                        example: 17
                      from:
                        type: string
                        example: null
                      last_page:
                        type: integer
                        example: 1
                      links:
                        type: array
                        example:
                          -
                            url: 'http://homzen.test/api/v1/facilities?page=16'
                            label: '&laquo; Previous'
                            active: false
                          -
                            url: 'http://homzen.test/api/v1/facilities?page=1'
                            label: '1'
                            active: false
                          -
                            url: null
                            label: 'Next &raquo;'
                            active: false
                        items:
                          type: object
                          properties:
                            url:
                              type: string
                              example: 'http://homzen.test/api/v1/facilities?page=16'
                            label:
                              type: string
                              example: '&laquo; Previous'
                            active:
                              type: boolean
                              example: false
                      path:
                        type: string
                        example: 'http://homzen.test/api/v1/facilities'
                      per_page:
                        type: integer
                        example: 17
                      to:
                        type: string
                        example: null
                      total:
                        type: integer
                        example: 11
                  error:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: null
      tags:
        - 'Real Estate'
      security: []
...